+++ packages/component-a/index.js
@@ -14,3 +14,10 @@
 
 
 // Fix for A
+// Break for A, B, C
+// Fix for A
+// Feat for A
+// Fix for A
+// Break for A, B, C
+// Feature for A, B
+// Fix for A, B, C