+++ packages/component-b/index.js
@@ -15,3 +15,8 @@
 // Feature for A, B, C
 
 // Fix for B
+// Fix for B
+// Break for A, B, C
+// Break for A, B, C
+// Feature for A, B
+// Fix for A, B, C