+++ packages/component-c/index.js
@@ -4,3 +4,7 @@
 
 // Feature for A, B, C
 // Feature for A, B, C
+// Break for A, B, C
+// Fix for C
+// Break for A, B, C
+// Fix for A, B, C