+++ packages/atlas-animation/CHANGELOG.md
@@ -1,10 +1,45 @@
-<a name="1.1.5"></a>
-## 1.1.5 (2016-06-06)
+<a name="1.1.6"></a>
+## [1.1.6](https://bitbucket.org/atlassian/[object Object]/compare/0.0.1-atlas-animation3...v1.1.6) (2016-06-14)
 
 
 ### Bug Fixes
 
-* spike webanimations-js polyfill and react integration ([653f208](https://bitbucket.org/atlassian/atlaskit-spike/commits/653f208))
+* **skate:** Update to Skate 1.0.0-beta.0 ([057edcb](https://bitbucket.org/atlassian/[object Object]/commits/057edcb))
+
+
+
+<a name="0.0.1-atlas-animation3"></a>
+## [0.0.1-atlas-animation3](https://bitbucket.org/atlassian/[object Object]/compare/1.1.6-atlas-animation...0.0.1-atlas-animation3) (2016-06-08)
+
+
+
+<a name="1.1.6-atlas-animation"></a>
+## [1.1.6-atlas-animation](https://bitbucket.org/atlassian/[object Object]/compare/1.1.6-semver-tag-for-atlas-animation...1.1.6-atlas-animation) (2016-06-08)
+
+
+
+<a name="1.1.6-semver-tag-for-atlas-animation"></a>
+## [1.1.6-semver-tag-for-atlas-animation](https://bitbucket.org/atlassian/[object Object]/compare/1.1.5-semver-tag-for-atlas-animation...1.1.6-semver-tag-for-atlas-animation) (2016-06-08)
+
+
+
+<a name="1.1.5-semver-tag-for-atlas-animation"></a>
+## [1.1.5-semver-tag-for-atlas-animation](https://bitbucket.org/atlassian/[object Object]/compare/1.1.0-semver-tag-for-atlas-animation...1.1.5-semver-tag-for-atlas-animation) (2016-06-06)
+
+
+### Bug Fixes
+
+* spike webanimations-js polyfill and react integration ([653f208](https://bitbucket.org/atlassian/[object Object]/commits/653f208))
+
+
+
+<a name="1.1.0-semver-tag-for-atlas-animation"></a>
+# [1.1.0-semver-tag-for-atlas-animation](https://bitbucket.org/atlassian/[object Object]/compare/1.0.0-semver-tag-for-atlas-animation...1.1.0-semver-tag-for-atlas-animation) (2016-06-05)
+
+
+
+<a name="1.0.0-semver-tag-for-atlas-animation"></a>
+# 1.0.0-semver-tag-for-atlas-animation (2016-06-02)