+++ packages/joshable-content/CHANGELOG.md
@@ -1,15 +1,25 @@
-<a name="11.2.0"></a>
-# 11.2.0 (2016-06-06)
+<a name="12.0.0"></a>
+# [12.0.0](https://bitbucket.org/atlassian/[object Object]/compare/0.0.1-atlas-animation3...v12.0.0) (2016-06-14)
+
+
+
+<a name="0.0.1-atlas-animation3"></a>
+## [0.0.1-atlas-animation3](https://bitbucket.org/atlassian/[object Object]/compare/1.1.6-atlas-animation...0.0.1-atlas-animation3) (2016-06-08)
+
+
+
+<a name="1.1.6-atlas-animation"></a>
+## [1.1.6-atlas-animation](https://bitbucket.org/atlassian/[object Object]/compare/34dcdd5...1.1.6-atlas-animation) (2016-06-08)
 
 
 ### Bug Fixes
 
-* **all:** bump ([34dcdd5](https://bitbucket.org/atlassian/atlaskit-spike/commits/34dcdd5))
+* **all:** bump ([34dcdd5](https://bitbucket.org/atlassian/[object Object]/commits/34dcdd5))
 
 
 ### Features
 
-* **test:** breaking change ([51f431d](https://bitbucket.org/atlassian/atlaskit-spike/commits/51f431d))
+* **test:** breaking change ([51f431d](https://bitbucket.org/atlassian/[object Object]/commits/51f431d))
 
 
 ### BREAKING CHANGES